requirements.txt
streamlit 
openai
validators
python-dotenv